// src/app/Models/editcustmoreDetail.1.ts
export class editcustmoreDetail {
    user_id: string;
    customer_id: string;
    last_name: string;
    first_name: string;
    kana_last_name: string;
    kana_first_name: string;
    service_id: string;
    email: string;
    tel: string;
    city: string;
    address1: string;
    address2: string;
    zip_code: string;
    gender: string;
    dob: string;
    age: string;
    job: string;
    subscription_of_news: string;
    know_about_company: string;
    how_did_you_come: string;
}