// src/app/Models/viewcustmoredetailResponse.model.ts
export class viewcustmoredetailResponse {

    id: string;
    user_id: string;
    salon_id: string;
    service_id: string;
    jts_customer_code: string;
    hotpaper_customer_code: string;
    customer_code: string;
    name: string;
    kana_first_name: string;
    kana_last_name: string;
    first_name: string;
    last_name: string;
    email: string;
    kana: string;
    gender: string;
    dob: string;
    age: string;
    tel: string;
    zip_code: string;
    prefecture: string;
    city: string;
    address1: string;
    address2: string;
    job: string;
    subscription_of_news: string;
    know_about_company: string;
    how_did_you_come: string;
    status: string;
    is_gmail: string;
    salon_board: string;
    verification_code: string;
    last_visited: string;
    created: string;
    modified: string;
    salon_name: string;
    service_name: string;

}